chore: move config submodules from startup module to config module
