+++ src/deadline_worker_agent/config/errors.py
@@ -1,0 +1,7 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+
+
+class ConfigurationError(Exception):
+    """An exception raised when an error is encountered loading configuration"""
+
+    pass
