+++ test/unit/aws/deadline/test_update_worker.py
@@ -13,14 +13,14 @@ from deadline_worker_agent.aws.deadline import (
 )
 
 import deadline_worker_agent.aws.deadline as deadline_mod
-from deadline_worker_agent.startup.config import Configuration
+from deadline_worker_agent.config import Configuration
 from deadline_worker_agent.api_models import (
     LogConfiguration,
     HostProperties,
     IpAddresses,
     WorkerStatus,
 )
-from deadline_worker_agent.startup.cli_args import ParsedCommandLineArguments
+from deadline_worker_agent.config.cli_args import ParsedCommandLineArguments
 from deadline_worker_agent.log_sync.cloudwatch import (
     LOG_CONFIG_OPTION_GROUP_NAME_KEY,
     LOG_CONFIG_OPTION_STREAM_NAME_KEY,
